+++ packages/frontend/routes/saksbilde/sykepengegrunnlag/inntekt/Refusjon/Refusjon.tsx
@@ -1,18 +1,18 @@
-import classNames from 'classnames';
+import { BeløpInput } from './BeløpInput';
 import dayjs from 'dayjs';
 import React, { useEffect } from 'react';
-import { Controller, FieldError, useFieldArray, useFormContext } from 'react-hook-form';
+import { Controller } from 'react-hook-form';
 
-import { CaseworkerFilled } from '@navikt/ds-icons';
-import { BodyShort, UNSAFE_DatePicker as DatePicker, Fieldset } from '@navikt/ds-react';
+import { UNSAFE_DatePicker as DatePicker, Fieldset } from '@navikt/ds-react';
 
-import { Button } from '@components/Button';
-import { Endringstrekant } from '@components/Endringstrekant';
-import { Flex } from '@components/Flex';
-import { Kilde } from '@components/Kilde';
-import { Kildetype } from '@io/graphql';
 import { Refusjonsopplysning } from '@io/http';
-import { ISO_DATOFORMAT, NORSK_DATOFORMAT } from '@utils/date';
+import { ISO_DATOFORMAT } from '@utils/date';
+
+import { DatoInput } from './DatoInput';
+import { KildeInput } from './KildeInput';
+import { LeggTilRefusjonsopplysningKnapp } from './LeggTilRefusjonsopplysningKnapp';
+import { SlettRefusjonsopplysningKnapp } from './SlettRefusjonsopplysningKnapp';
+import { useRefusjonFormField } from './useRefusjonFormField';
 
 import styles from './Refusjon.module.css';
 
@@ -206,173 +206,3 @@ export const Refusjon = ({ fraRefusjonsopplysninger, lokaleRefusjonsopplysninger
         </Fieldset>
     );
 };
-
-interface DatoInputProps {
-    updateDato: (dato: Maybe<string>) => void;
-    clearError: () => void;
-    dato?: Maybe<string>;
-    error?: FieldError;
-}
-
-const DatoInput = ({ updateDato, clearError, dato, error }: DatoInputProps) => (
-    <DatePicker.Input
-        label=""
-        className={styles.DateInput}
-        size="small"
-        placeholder="dd.mm.åååå"
-        onBlur={(e) => {
-            const nyDato = dayjs(e.target.value, NORSK_DATOFORMAT).isValid()
-                ? dayjs(e.target.value, NORSK_DATOFORMAT).format(ISO_DATOFORMAT)
-                : e.target.value === ''
-                ? null
-                : e.target.value;
-            if (nyDato === dato) return;
-
-            clearError();
-            updateDato(nyDato);
-        }}
-        defaultValue={
-            dato && dayjs(dato, ISO_DATOFORMAT).isValid()
-                ? dayjs(dato, ISO_DATOFORMAT)?.format(NORSK_DATOFORMAT)
-                : dato ?? undefined
-        }
-        error={error?.message}
-    />
-);
-
-interface BeløpInputProps {
-    beløp: number;
-    updateBeløp: (beløp: number) => void;
-    clearError: () => void;
-    error?: FieldError;
-}
-
-const BeløpInput = ({ beløp, updateBeløp, clearError, error }: BeløpInputProps) => (
-    <input
-        className={classNames({
-            [styles.BeløpInput]: true,
-            [styles.InputError]: error?.message,
-        })}
-        type="number"
-        onBlur={(event) => {
-            const nyttBeløp = Number(event.target.value);
-            if (nyttBeløp === beløp) return;
-            clearError();
-            updateBeløp(Number(event.target.value));
-        }}
-        defaultValue={beløp && Math.round((beløp + Number.EPSILON) * 100) / 100}
-    />
-);
-
-interface KildeInputProps {
-    kilde: string;
-    lokalRefusjonsopplysning?: Refusjonsopplysning;
-    fraRefusjonsopplysning?: Refusjonsopplysning;
-}
-
-const KildeInput = ({ kilde, lokalRefusjonsopplysning, fraRefusjonsopplysning }: KildeInputProps) => (
-    <Flex alignItems="center">
-        {kilde === Kildetype.Inntektsmelding && (
-            <Kilde type={kilde} className={styles.Ikon}>
-                IM
-            </Kilde>
-        )}
-        {kilde === Kildetype.Saksbehandler &&
-            (lokalRefusjonsopplysning && !erLikRefusjonsopplysning(lokalRefusjonsopplysning, fraRefusjonsopplysning) ? (
-                <div style={{ position: 'relative', width: '20px' }}>
-                    <Endringstrekant />
-                </div>
-            ) : (
-                <Kilde type={kilde} className={styles.Ikon}>
-                    <CaseworkerFilled title="Caseworker-ikon" height={12} width={12} />
-                </Kilde>
-            ))}
-    </Flex>
-);
-
-const erLikRefusjonsopplysning = (a?: Refusjonsopplysning, b?: Refusjonsopplysning) =>
-    JSON.stringify(a) === JSON.stringify(b);
-
-interface LeggTilRefusjonsopplysningKnappProps {
-    onClick: () => void;
-}
-
-const LeggTilRefusjonsopplysningKnapp = ({ onClick }: LeggTilRefusjonsopplysningKnappProps) => (
-    <Button type="button" onClick={onClick} className={styles.Button} style={{ marginTop: 0 }}>
-        <BodyShort>+ Legg til</BodyShort>
-    </Button>
-);
-
-interface SlettRefusjonsopplysningKnappProps {
-    onClick: () => void;
-}
-
-const SlettRefusjonsopplysningKnapp = ({ onClick }: SlettRefusjonsopplysningKnappProps) => (
-    <Button type="button" onClick={onClick} className={styles.Button} style={{ justifySelf: 'flex-end' }}>
-        <BodyShort>Slett</BodyShort>
-    </Button>
-);
-
-interface RefusjonFormValues {
-    name: string;
-    refusjonsopplysninger: { fom: string; tom?: Maybe<string>; beløp: number; kilde: string }[];
-}
-
-function useRefusjonFormField() {
-    const { formState, control, clearErrors } = useFormContext<RefusjonFormValues>();
-
-    const { fields, append, remove, replace, update } = useFieldArray<RefusjonFormValues>({
-        control,
-        name: 'refusjonsopplysninger',
-    });
-
-    const addRefusjonsopplysning = () => {
-        append({
-            fom: '',
-            tom: null,
-            beløp: 0,
-            kilde: Kildetype.Saksbehandler,
-        });
-    };
-
-    const removeRefusjonsopplysning = (index: number) => () => {
-        remove(index);
-    };
-
-    const replaceRefusjonsopplysninger = (refusjonsopplysninger: Refusjonsopplysning[]) => {
-        replace(
-            refusjonsopplysninger
-                .sort(
-                    (a: Refusjonsopplysning, b: Refusjonsopplysning) =>
-                        new Date(b.fom).getTime() - new Date(a.fom).getTime(),
-                )
-                .map((refusjonsopplysning) => {
-                    return {
-                        ...refusjonsopplysning,
-                        beløp: Math.round((refusjonsopplysning.beløp + Number.EPSILON) * 100) / 100,
-                    };
-                }),
-        );
-    };
-
-    const updateRefusjonsopplysninger = (
-        fom: string,
-        tom: Maybe<string>,
-        beløp: number,
-        index: number,
-        kilde = Kildetype.Saksbehandler,
-    ) => {
-        update(index, { fom, tom, beløp, kilde });
-    };
-
-    return {
-        fields,
-        control,
-        clearErrors,
-        formState,
-        addRefusjonsopplysning,
-        removeRefusjonsopplysning,
-        replaceRefusjonsopplysninger,
-        updateRefusjonsopplysninger,
-    };
-}
